backend/requirements.txt
fastapi[standard]
pydantic
pytest
pytest-asyncio
httpx
python-dotenv
transformers
torch
langchain
langchain-community
langchain-core
langchain-ollama